+int64

+++ C-fast_longlong_64bit/fft.h
@@ -2,8 +2,8 @@
 #define INTMULSAMPLE (55-4)
 
 struct _sample {
-    long i;
-    long q;
+    long long i;
+    long long q;
 };
 // step: N*log2(N)
 void fft     (int log2point, struct _sample *restrict xy_out, const struct _sample *restrict xy_in);
